test if ~/.cache/yt-dlp exists before rm

+++ entrypoint.sh
@@ -1,4 +1,6 @@
 #! /bin/sh
-rm -rf ~/.cache/yt-dlp
+if [ -d "~/.cache/yt-dlp" ]; then
+  rm -rf ~/.cache/yt-dlp
+fi
 cd /app
 /app/inv_sig_helper_rust
